refactor: ConsoleInput 추가 및 GameConsole 수정
